optimize, restructure, get rid of fgets
time to write about historical baggage.

let's say you were to write a tool that allows you to view lines from a file
based on a line number, so that you can easily check what's on a specific line
on a text or source code file or whatever. you'll need some function to
actually *read* the lines from the file.

in POSIXland, you've got getline, which relies on memory allocation and is fine
until someone throws in a 16 GB one-liner at you and then complains how all
memory allocation and no printing makes using the tool a complete waste of time.

but huzzah, along comes the C standard library with this thing called "fgets".
you give it a buffer, its size and the file, and it reads you a line of text
into that buffer. excellent! exactly what you need. it even tells you if there
was an EOF or I/O error.

what if the line is too long now? fgets will simply not read past the buffer,
so it'll read a part of the line and then keep going the next time around.
the way fgets works is obviously catered towards how C strings work. so no
matter what, the buffer will have a null terminator at the very end.

now here's a question for you; how do you determine if fgets read in a complete
line of text or not? checking the actual end of the buffer is no good -- what
if we read in a really long line and a short line afterwards?

what about using ftell? no dice, since what if we're piping from stdin?
ftell will just give us -1.

well, there's a null terminator. why not use strlen? we can use that to figure
out where the line-ending null terminator is and thus find the "end" of the
data we read.

yet that is worth nothing when the files you deal with can contain null bytes.
sure, if it were up to someone more ambitious, they'd probably push for
making it a crime to have null bytes in text files, but I'm not the one making
the decisions on what files people are going to use this tool for.
thus, the tool needs to be able to gracefully handle null bytes in files and
we cannot rely on the null byte *in any way*. it's not like fgets stops
whenever it reads a null character either, so why should we?

so okay, another idea. we'll try to find the newline character from the buffer,
which is something memchr is perfectly suited for. except now we have to scan
through the buffer *twice*, the first time around to actually read it (as
fgets does) and the second time to figure out the length of the line. good
luck optimizing that!

wait, someone found a bigger problem with this approach. if we have a text file
ending in a line that does not have a newline character, we have *basically no
way* to know its true length. so, imagine this perfect storm: piping a text
file through stdin in which the final line does not have a newline but contains
null terminator bytes (again, I don't make the rules). now, if you use fgets,
you simply CANNOT tell in any reliable way how long the line you read actually
was. you cannot use ftell because of the pipe. you cannot use strlen because
the line might have null bytes in it. you cannot scan for a newline because
THERE ISN'T ONE. the only possible option is to maybe somehow fill the buffer
with -1 beforehand and then scan backwards for a null byte if we reach eof...
but good luck making that approach actually work with an acceptable level of
performance and it cannot handle all test cases (what if the file ends with
one or more -1 bytes?)

congratulations, fgets is broken. all of this could be solved if fgets would
simply return some information on how many bytes it read, be it a size_t or
ssize_t or int or whatever or a pointer to the end of the line (the "null
terminator" it places). this would require basically NO extra work from the
function because it must know where it'll be writing anyway.

so this is where we finally get to start talking about historical baggage.
strike one; fgets returns a pointer. this is (almost certainly) because it's
based on gets. this is not a problem if the pointer would point to the end of
the data we read. ...strike two; fgets returns a pointer *that is just a copy
of the pointer to the buffer given as a parameter*. this is (almost certainly)
because it's based on gets. this would have been a perfect opportunity to make
the return value of fgets actually make sense, but no. instead just parrot and
base it on a function that is so unsafe IT WAS REMOVED FROM THE C STANDARD!

so strike three; gets (may I be forgiven for uttering this word three times)
returns a pointer that is just a copy of the pointer to the buffer given as
a parameter. why? nobody knows. someone (perhaps even Ritchie himself) made
that decision a long time ago and we still have to live with its consequences.

so what are the alternatives? you can find the newlines yourself and just
read stuff into a buffer. POSIX read is great for this, fread not so much
because it is too hardheaded to stop reading until there's an EOF or the end of
buffer is reached. in other words, screw the terminal and pipes, keep giving me
more. I'm sure there's some great historical reasons for that as well.

sure, all this is minor compared to some other trainwrecks (locale system),
but it still further adds to the pile of things that suck in coding C.